Fix hexadecimal value conversion on HexIntegerField

+++ push_notifications/fields.py
@@ -60,7 +60,7 @@ class HexIntegerField(six.with_metaclass(models.SubfieldBase, models.BigIntegerF
 		if value is None or value == "":
 			return None
 		if isinstance(value, six.string_types):
-			value = int(value, 16)
+			value = long(value, 16)
 		# on postgres only, interpret as signed
 		if connection.settings_dict["ENGINE"] in postgres_engines:
 			value = struct.unpack("q", struct.pack("Q", value))[0]
